Late-night janitorial work. Cleaned up and added some more documentation.

+++ src/main/java/com/revature/projectzero/services/UserCoursesService.java
@@ -4,16 +4,17 @@ import com.revature.projectzero.documents.UserCourses;
 import com.revature.projectzero.repositories.UserCoursesRepository;
 import com.revature.projectzero.util.UserSession;
 import com.revature.projectzero.util.exceptions.AlreadyRegisteredForCourseException;
+import com.revature.projectzero.util.exceptions.CourseNotJoinedException;
 import com.revature.projectzero.util.exceptions.NoCoursesJoinedException;
 import com.revature.projectzero.util.exceptions.NotRegisteredForCourseException;
 
 import java.util.List;
 
+// Service for handling most user course list business logic and passing information into the UserCourse repository
 
 public class UserCoursesService {
 
     private final UserCoursesRepository userCourseListRepo;
-
     private final UserSession session;
 
     public UserCoursesService(UserCoursesRepository userCourseRepo, UserSession session) {
@@ -21,13 +22,14 @@ public class UserCoursesService {
         this.session = session;
     }
 
+    // Initialize a user's course list on the database when they register
     public void initialize(UserCourses newUserCourseList){
         userCourseListRepo.save(newUserCourseList);
     }
 
+    // Checks to see if the user has already joined a course, passes the course requested and the username to the Repo if not
     public void joinCourse(String courseToJoin){
 
-        // Grabbing values to iterate over here instead of on one line so it looks nicer and is easier to follow
         String un = session.getCurrentUser().getUsername();
         List<String> userCourses = userCourseListRepo.findRegisteredCoursesByUsername(un);
 
@@ -46,6 +48,22 @@ public class UserCoursesService {
     }
 
 
+    // Compares user's entered course against a list of the user's courses, ignores case.
+    // Returns proper case if the course is found.
+    public String verifyCourseEntry(List<String> joinedCourses, String userEntry){
+
+        for (String course:joinedCourses)
+        {
+            if (course.equalsIgnoreCase(userEntry))
+                return course;
+        }
+
+        System.out.println("You have not registered for a course by this name.");
+        throw new NotRegisteredForCourseException("User attempted to withdraw for a course they were not enrolled in.");
+
+    }
+
+    // Verifies if a user is in any courses before attempting to remove the course.
     public void leaveCourse(String courseToLeave) {
 
         String un = session.getCurrentUser().getUsername();
@@ -53,22 +71,17 @@ public class UserCoursesService {
 
         // Should not actually be reached, but keeping for posterity
         if (userCourses.isEmpty())
-            throw new NoCoursesJoinedException("You have not registered for any courses!");
-
-        userCourseListRepo.removeCourseFromUserList(courseToLeave,un);
-
-    }
-
-    public String verifyCourseEntry(List<String> joinedCourses, String userEntry){
-
-        for (String course:joinedCourses)
         {
-            if (course.equalsIgnoreCase(userEntry.toLowerCase()))
-                return course;
+            System.out.println("You have not registered for any courses!");
+            throw new NoCoursesJoinedException("The user has not registered for any courses.");
+        }
+        if(!userCourses.contains(courseToLeave))
+        {
+            System.out.println("You have not joined that course!");
+            throw new CourseNotJoinedException("User attempted to leave a course they have not joined.");
         }
 
-        System.out.println("You have not registered for a course by this name.");
-        throw new NotRegisteredForCourseException("User attempted to withdraw for a course they were not enrolled in.");
+        userCourseListRepo.removeCourseFromUserList(courseToLeave,un);
 
     }
